+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 ### Effect DSL
 - Missing brackets are now reported with less misleading error messages.
 - Improved error messages for missing semicolons.
+- Added error messages for missing commas in DSL expressions.
+
 
 ## tachyonfx 0.13.0 - 2025-03-30